- Added code to preserve permissions of invoices folder in the installer. - Fixed a problem with german umlauts in monitor. - Added code to update.php and autoupdate.php that checks the mysql root password before the update starts. - Fixed: FS#1848 - changing password of ssh user >>> rights in webXX are changed
git-svn-id: svn://svn.ispconfig.org/ispconfig3/branches/ispconfig-3.0.4@2778 c1a6f2c7-8504-0410-933a-a4dae0184b0a

+++ server/plugins-available/shelluser_base_plugin.inc.php
@@ -235,7 +235,7 @@ class shelluser_base_plugin {
 		if (!file_exists($sshkeys)){
 			// add root's key
 			$app->file->mkdirs($sshdir, '0755');
-			file_put_contents($sshkeys, file_get_contents('/root/.ssh/authorized_keys'));
+			if(is_file('/root/.ssh/authorized_keys')) file_put_contents($sshkeys, file_get_contents('/root/.ssh/authorized_keys'));
 		
 			// Remove duplicate keys
 			$existing_keys = file($sshkeys);
